Entities can now wander
This does interfere with EntityPlayerResponse, so be warned. I am going to be updating response to comment it out, but if you are wondering why, it is because it interferes with this script. Specifically, it makes the entities go through the flooring because it is trying to look at the player with eye to eye contact and if the player is taller, the entity rotates too much, and then tries to fly.