Made drops start being dropped from objects
Instead of just destroying an object when an object is destroyed, it will now actually start dropping the drops. ObjectFunctions.cs is going to house a lot of things, so there is a lot more work to go, so I also added a ToDo at the top. At the moment there is not much, only to make the objects spread out a bit more so everything is not on top of each other, as well as have a chance to have rare drops. Currently, it is not too major because a lot of the objects in the game are basic and not "rare", but eventually there might be more objects. Like an arrow might have a broken fragment or something.